+++ tests/ByteDev.Json.SystemTextJson.UnitTests/TestEntities/TestGuidEntity.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Text.Json.Serialization;
+
+namespace ByteDev.Json.SystemTextJson.UnitTests.TestEntities
+{
+    public class TestGuidEntity
+    {
+        [JsonPropertyName("myGuid")]
+        public Guid MyGuid { get; set; }
+    }
+}